feature-77: Added validation to the building controller

+++ src/controllers/buildingsController.js
@@ -35,6 +35,28 @@ const createBuilding = async (req, res) => {
       msg: 'Error: Missing required fields to create a building'
     });
   }
+  const boilers = await models.Boilers.find({ _id: req.body.boilers });
+  console.log(boilers);
+  const missingBoilers = [];
+  req.body.boilers.forEach(boiler => {
+    // eslint-disable-next-line no-underscore-dangle
+    const found = boilers.find(existingBoiler => existingBoiler._id.toString() === boiler);
+    if (!found) {
+      missingBoilers.push(boiler);
+    }
+  });
+  if (missingBoilers.length > 0) {
+    return res.status(500).json({
+      msg: 'The next boilers were not found in the database',
+      missingBoilers,
+    });
+  }
+  const company = await models.Companies.findById(req.body.company);
+  if (!company) {
+    return res.status(500).json({
+      msg: 'The company assigned to the building was not found in the database.'
+    });
+  }
 
   const building = new models.Building({
     name: req.body.name,
@@ -62,6 +84,28 @@ const updateBuilding = async (req, res) => {
     });
   }
   try {
+    const boilers = await models.Boilers.find({ _id: req.body.boilers });
+    console.log(boilers);
+    const missingBoilers = [];
+    req.body.boilers.forEach(boiler => {
+      // eslint-disable-next-line no-underscore-dangle
+      const found = boilers.find(existingBoiler => existingBoiler._id.toString() === boiler);
+      if (!found) {
+        missingBoilers.push(boiler);
+      }
+    });
+    if (missingBoilers.length > 0) {
+      return res.status(500).json({
+        msg: 'The next boilers were not found in the database',
+        missingBoilers,
+      });
+    }
+    const company = await models.Companies.findById(req.body.company);
+    if (!company) {
+      return res.status(500).json({
+        msg: 'The company assigned to the building was not found in the database.'
+      });
+    }
     const result = await models.Building.findByIdAndUpdate(req.params.id, req.body, { new: true, });
 
     if (!result) {
@@ -79,6 +123,12 @@ const updateBuilding = async (req, res) => {
 
 const deleteBuilding = async (req, res) => {
   try {
+    const appointment = await models.Appointments.find({ building: req.params.id });
+    if (appointment) {
+      return res.status(500).json({
+        msg: 'The building you\'re trying to deleted has an appointment',
+      });
+    }
     const result = await models.Building.findByIdAndDelete(req.params.id);
 
     if (!result) {
